requirements.txt
# Bibliotecas principais
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
pillow>=8.2.0

# Processamento geoespacial
rasterio>=1.2.0
geopandas>=0.9.0
shapely>=1.7.0
rasterstats>=0.15.0
pyproj>=3.0.0

# API e Web
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
python-dotenv>=0.19.0

# Supabase
supabase>=0.0.8

# Geração de relatórios
reportlab>=3.6.0
